Minor: moves /utils/sendbird.ts to /src/app/gigs/chat and updates imports
